requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
black==20.8b1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-environ==0.4.5
django-mssql-backend==2.8.1
djangorestframework==3.11.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==5.5.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
packaging==20.3
pathspec==0.8.0
pep517==0.8.2
progress==1.5
pylint==2.6.0
pyodbc==4.0.30
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
regex==2020.7.14
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
pymssql
